RMS v2 Initial Update
Version 2.0.0 of RMS. This is designed to be totally incompatible w/ the v1 version. This does not mean we'll stop supporting v1. Consult the webpage (will update early next mo (September)) for specific information.
This is a very simple implementation of the features in v1. I want to make the project more streamlined and follow our design purpose - as a clean one-liner approach to Scikit-learn. It is sure that my main development of the RMS package thru 2025 Q3/4 and even maybe onto 26Q1 will be v2. My current plan is to fully drop v1 support (this means that v1 will NO LONGER RECEIVE UPDATES) late 2025 or early 2026, perhaps sometime in Dec or Jan.
I think it's reasonable to not write too much release notes... so I will update/deploy the web page in the following days (as well as the index,changelog whatever) so more information can be seen there.
Shall there be any questions, feel free to open an issue!